+++ packages/firebase_auth/android/src/main/java/io/flutter/plugins/firebaseauth/FirebaseAuthPlugin.java
@@ -29,7 +29,6 @@ import java.util.concurrent.TimeUnit;
 /** Flutter plugin for Firebase Auth. */
 public class FirebaseAuthPlugin implements MethodCallHandler {
   private final PluginRegistry.Registrar registrar;
-  private final FirebaseAuth firebaseAuth;
   private final SparseArray<FirebaseAuth.AuthStateListener> authStateListeners =
       new SparseArray<>();
   private final SparseArray<PhoneAuthProvider.ForceResendingToken> forceResendingTokens =
@@ -51,86 +50,98 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     this.registrar = registrar;
     this.channel = channel;
     FirebaseApp.initializeApp(registrar.context());
-    this.firebaseAuth = FirebaseAuth.getInstance();
+  }
+
+  private FirebaseAuth getAuth(MethodCall call) {
+    Map<String, Object> arguments = (Map<String, Object>) call.arguments;
+    String appName = (String) arguments.get("app");
+    FirebaseApp app = FirebaseApp.getInstance(appName);
+    return FirebaseAuth.getInstance(app);
   }
 
   @Override
   public void onMethodCall(MethodCall call, Result result) {
     switch (call.method) {
       case "currentUser":
-        handleCurrentUser(call, result);
+        handleCurrentUser(call, result, getAuth(call));
         break;
       case "signInAnonymously":
-        handleSignInAnonymously(call, result);
+        handleSignInAnonymously(call, result, getAuth(call));
         break;
       case "createUserWithEmailAndPassword":
-        handleCreateUserWithEmailAndPassword(call, result);
+        handleCreateUserWithEmailAndPassword(call, result, getAuth(call));
         break;
       case "fetchProvidersForEmail":
-        handleFetchProvidersForEmail(call, result);
+        handleFetchProvidersForEmail(call, result, getAuth(call));
         break;
       case "sendPasswordResetEmail":
-        handleSendPasswordResetEmail(call, result);
+        handleSendPasswordResetEmail(call, result, getAuth(call));
         break;
       case "sendEmailVerification":
-        handleSendEmailVerification(call, result);
+        handleSendEmailVerification(call, result, getAuth(call));
         break;
       case "reload":
-        handleReload(call, result);
+        handleReload(call, result, getAuth(call));
         break;
       case "delete":
-        handleDelete(call, result);
+        handleDelete(call, result, getAuth(call));
         break;
       case "signInWithEmailAndPassword":
-        handleSignInWithEmailAndPassword(call, result);
+        handleSignInWithEmailAndPassword(call, result, getAuth(call));
         break;
       case "signInWithGoogle":
-        handleSignInWithGoogle(call, result);
+        handleSignInWithGoogle(call, result, getAuth(call));
         break;
       case "signInWithCustomToken":
-        handleSignInWithCustomToken(call, result);
+        handleSignInWithCustomToken(call, result, getAuth(call));
         break;
       case "signInWithFacebook":
-        handleSignInWithFacebook(call, result);
+        handleSignInWithFacebook(call, result, getAuth(call));
         break;
       case "signInWithTwitter":
-        handleSignInWithTwitter(call, result);
+        handleSignInWithTwitter(call, result, getAuth(call));
         break;
       case "signOut":
-        handleSignOut(call, result);
+        handleSignOut(call, result, getAuth(call));
         break;
       case "getIdToken":
-        handleGetToken(call, result);
+        handleGetToken(call, result, getAuth(call));
         break;
       case "linkWithEmailAndPassword":
-        handleLinkWithEmailAndPassword(call, result);
+        handleLinkWithEmailAndPassword(call, result, getAuth(call));
         break;
       case "linkWithGoogleCredential":
-        handleLinkWithGoogleCredential(call, result);
+        handleLinkWithGoogleCredential(call, result, getAuth(call));
         break;
       case "linkWithFacebookCredential":
-        handleLinkWithFacebookCredential(call, result);
+        handleLinkWithFacebookCredential(call, result, getAuth(call));
         break;
-      case "updateProfile":
-        handleUpdateProfile(call, result);
+      case "linkWithTwitterCredential":
+        handleLinkWithTwitterCredential(call, result, getAuth(call));
         break;
       case "updateEmail":
-        handleUpdateEmail(call, result);
+        handleUpdateEmail(call, result, getAuth(call));
+        break;
+      case "updatePassword":
+        handleUpdatePassword(call, result, getAuth(call));
+        break;
+      case "updateProfile":
+        handleUpdateProfile(call, result, getAuth(call));
         break;
       case "startListeningAuthState":
-        handleStartListeningAuthState(call, result);
+        handleStartListeningAuthState(call, result, getAuth(call));
         break;
       case "stopListeningAuthState":
-        handleStopListeningAuthState(call, result);
+        handleStopListeningAuthState(call, result, getAuth(call));
         break;
       case "verifyPhoneNumber":
-        handleVerifyPhoneNumber(call, result);
+        handleVerifyPhoneNumber(call, result, getAuth(call));
         break;
       case "signInWithPhoneNumber":
-        handleSignInWithPhoneNumber(call, result);
+        handleSignInWithPhoneNumber(call, result, getAuth(call));
         break;
       case "setLanguageCode":
-        handleSetLanguageCode(call, result);
+        handleSetLanguageCode(call, result, getAuth(call));
         break;
       default:
         result.notImplemented();
@@ -138,7 +149,9 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     }
   }
 
-  private void handleSignInWithPhoneNumber(MethodCall call, Result result) {
+  private void handleSignInWithPhoneNumber(
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String verificationId = arguments.get("verificationId");
     String smsCode = arguments.get("smsCode");
@@ -150,7 +163,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleVerifyPhoneNumber(MethodCall call, Result result) {
+  private void handleVerifyPhoneNumber(
+      MethodCall call, Result result, final FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     final int handle = call.argument("handle");
     String phoneNumber = call.argument("phoneNumber");
@@ -243,7 +257,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     return exceptionMap;
   }
 
-  private void handleLinkWithEmailAndPassword(MethodCall call, Result result) {
+  private void handleLinkWithEmailAndPassword(
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String email = arguments.get("email");
@@ -256,26 +271,23 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleCurrentUser(MethodCall call, final Result result) {
-    final FirebaseAuth.AuthStateListener listener =
-        new FirebaseAuth.AuthStateListener() {
-          @Override
-          public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
-            firebaseAuth.removeAuthStateListener(this);
-            FirebaseUser user = firebaseAuth.getCurrentUser();
-            Map<String, Object> userMap = mapFromUser(user);
-            result.success(userMap);
-          }
-        };
-
-    firebaseAuth.addAuthStateListener(listener);
+  private void handleCurrentUser(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+    FirebaseUser user = firebaseAuth.getCurrentUser();
+    if (user == null) {
+      result.success(null);
+      return;
+    }
+    Map<String, Object> userMap = mapFromUser(user);
+    result.success(userMap);
   }
 
-  private void handleSignInAnonymously(MethodCall call, final Result result) {
+  private void handleSignInAnonymously(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth.signInAnonymously().addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleCreateUserWithEmailAndPassword(MethodCall call, final Result result) {
+  private void handleCreateUserWithEmailAndPassword(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String email = arguments.get("email");
@@ -286,7 +298,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleFetchProvidersForEmail(MethodCall call, final Result result) {
+  private void handleFetchProvidersForEmail(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String email = arguments.get("email");
@@ -296,7 +309,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new ProvidersCompleteListener(result));
   }
 
-  private void handleSendPasswordResetEmail(MethodCall call, final Result result) {
+  private void handleSendPasswordResetEmail(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String email = arguments.get("email");
@@ -306,28 +320,23 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleSendEmailVerification(MethodCall call, final Result result) {
+  private void handleSendEmailVerification(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth
         .getCurrentUser()
         .sendEmailVerification()
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleReload(MethodCall call, final Result result) {
+  private void handleReload(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth
         .getCurrentUser()
         .reload()
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleDelete(MethodCall call, final Result result) {
-    firebaseAuth
-        .getCurrentUser()
-        .delete()
-        .addOnCompleteListener(new TaskVoidCompleteListener(result));
-  }
-
-  private void handleSignInWithEmailAndPassword(MethodCall call, final Result result) {
+  private void handleSignInWithEmailAndPassword(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String email = arguments.get("email");
@@ -338,7 +347,15 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleSignInWithGoogle(MethodCall call, final Result result) {
+  private void handleDelete(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+    firebaseAuth
+        .getCurrentUser()
+        .delete()
+        .addOnCompleteListener(new TaskVoidCompleteListener(result));
+  }
+
+  private void handleSignInWithGoogle(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String idToken = arguments.get("idToken");
@@ -349,7 +366,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleLinkWithGoogleCredential(MethodCall call, final Result result) {
+  private void handleLinkWithGoogleCredential(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String idToken = arguments.get("idToken");
@@ -361,7 +379,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleLinkWithFacebookCredential(MethodCall call, final Result result) {
+  private void handleLinkWithFacebookCredential(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String accessToken = arguments.get("accessToken");
@@ -372,7 +391,21 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleSignInWithFacebook(MethodCall call, final Result result) {
+  private void handleLinkWithTwitterCredential(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+    @SuppressWarnings("unchecked")
+    Map<String, String> arguments = (Map<String, String>) call.arguments;
+    String authToken = arguments.get("authToken");
+    String authTokenSecret = arguments.get("authTokenSecret");
+    AuthCredential credential = TwitterAuthProvider.getCredential(authToken, authTokenSecret);
+    firebaseAuth
+        .getCurrentUser()
+        .linkWithCredential(credential)
+        .addOnCompleteListener(new SignInCompleteListener(result));
+  }
+
+  private void handleSignInWithFacebook(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String accessToken = arguments.get("accessToken");
@@ -382,7 +415,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleSignInWithTwitter(MethodCall call, final Result result) {
+  private void handleSignInWithTwitter(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     String authToken = call.argument("authToken");
     String authTokenSecret = call.argument("authTokenSecret");
     AuthCredential credential = TwitterAuthProvider.getCredential(authToken, authTokenSecret);
@@ -391,7 +425,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleSignInWithCustomToken(MethodCall call, final Result result) {
+  private void handleSignInWithCustomToken(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     Map<String, String> arguments = call.arguments();
     String token = arguments.get("token");
     firebaseAuth
@@ -399,12 +434,12 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleSignOut(MethodCall call, final Result result) {
+  private void handleSignOut(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth.signOut();
     result.success(null);
   }
 
-  private void handleGetToken(MethodCall call, final Result result) {
+  private void handleGetToken(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, Boolean> arguments = (Map<String, Boolean>) call.arguments;
     boolean refresh = arguments.get("refresh");
@@ -424,7 +459,27 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
             });
   }
 
-  private void handleUpdateProfile(MethodCall call, final Result result) {
+  private void handleUpdateEmail(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+    @SuppressWarnings("unchecked")
+    Map<String, String> arguments = (Map<String, String>) call.arguments;
+    firebaseAuth
+        .getCurrentUser()
+        .updateEmail(arguments.get("email"))
+        .addOnCompleteListener(new TaskVoidCompleteListener(result));
+  }
+
+  private void handleUpdatePassword(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+    @SuppressWarnings("unchecked")
+    Map<String, String> arguments = (Map<String, String>) call.arguments;
+    firebaseAuth
+        .getCurrentUser()
+        .updatePassword(arguments.get("password"))
+        .addOnCompleteListener(new TaskVoidCompleteListener(result));
+  }
+
+  private void handleUpdateProfile(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
 
@@ -439,43 +494,11 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     firebaseAuth
         .getCurrentUser()
         .updateProfile(builder.build())
-        .addOnCompleteListener(
-            new OnCompleteListener<Void>() {
-              @Override
-              public void onComplete(@NonNull Task<Void> task) {
-                if (!task.isSuccessful()) {
-                  Exception e = task.getException();
-                  result.error(ERROR_REASON_EXCEPTION, e.getMessage(), null);
-                } else {
-                  result.success(null);
-                }
-              }
-            });
+        .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleUpdateEmail(MethodCall call, final Result result) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String email = arguments.get("email");
-
-    firebaseAuth
-        .getCurrentUser()
-        .updateEmail(email)
-        .addOnCompleteListener(
-            new OnCompleteListener<Void>() {
-              @Override
-              public void onComplete(@NonNull Task<Void> task) {
-                if (!task.isSuccessful()) {
-                  Exception e = task.getException();
-                  result.error(ERROR_REASON_EXCEPTION, e.getMessage(), null);
-                } else {
-                  result.success(null);
-                }
-              }
-            });
-  }
-
-  private void handleStartListeningAuthState(MethodCall call, final Result result) {
+  private void handleStartListeningAuthState(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     final int handle = nextHandle++;
     FirebaseAuth.AuthStateListener listener =
         new FirebaseAuth.AuthStateListener() {
@@ -492,18 +515,19 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
             channel.invokeMethod("onAuthStateChanged", Collections.unmodifiableMap(map));
           }
         };
-    FirebaseAuth.getInstance().addAuthStateListener(listener);
+    firebaseAuth.addAuthStateListener(listener);
     authStateListeners.append(handle, listener);
     result.success(handle);
   }
 
-  private void handleStopListeningAuthState(MethodCall call, final Result result) {
+  private void handleStopListeningAuthState(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     Map<String, Integer> arguments = call.arguments();
     Integer id = arguments.get("id");
 
     FirebaseAuth.AuthStateListener listener = authStateListeners.get(id);
     if (listener != null) {
-      FirebaseAuth.getInstance().removeAuthStateListener(listener);
+      firebaseAuth.removeAuthStateListener(listener);
       authStateListeners.remove(id);
       result.success(null);
     } else {
@@ -514,7 +538,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     }
   }
 
-  private void handleSetLanguageCode(MethodCall call, final Result result) {
+  private void handleSetLanguageCode(
+      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     @SuppressWarnings("unchecked")
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String language = arguments.get("language");
@@ -610,6 +635,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         providerData.add(Collections.unmodifiableMap(userInfoToMap(userInfo)));
       }
       Map<String, Object> userMap = userInfoToMap(user);
+      userMap.put("creationTimestamp", user.getMetadata().getCreationTimestamp());
+      userMap.put("lastSignInTimestamp", user.getMetadata().getLastSignInTimestamp());
       userMap.put("isAnonymous", user.isAnonymous());
       userMap.put("isEmailVerified", user.isEmailVerified());
       userMap.put("providerData", Collections.unmodifiableList(providerData));