+++ packages/google_maps_flutter/lib/src/location.dart
@@ -6,12 +6,6 @@ part of google_maps_flutter;
 
 /// A pair of latitude and longitude coordinates, stored as degrees.
 class LatLng {
-  /// The latitude in degrees between -90.0 and 90.0, both inclusive.
-  final double latitude;
-
-  /// The longitude in degrees between -180.0 (inclusive) and 180.0 (exclusive).
-  final double longitude;
-
   /// Creates a geographical location specified in degrees [latitude] and
   /// [longitude].
   ///
@@ -25,6 +19,12 @@ class LatLng {
         latitude =
             (latitude < -90.0 ? -90.0 : (90.0 < latitude ? 90.0 : latitude)),
         longitude = (longitude + 180.0) % 360.0 - 180.0;
+
+  /// The latitude in degrees between -90.0 and 90.0, both inclusive.
+  final double latitude;
+
+  /// The longitude in degrees between -180.0 (inclusive) and 180.0 (exclusive).
+  final double longitude;
 
   dynamic _toJson() {
     return <double>[latitude, longitude];
@@ -60,12 +60,6 @@ class LatLng {
 /// * lng ∈ [-180, `northeast.longitude`] ∪ [`southwest.longitude`, 180[,
 ///   if `northeast.longitude` < `southwest.longitude`
 class LatLngBounds {
-  /// The southwest corner of the rectangle.
-  final LatLng southwest;
-
-  /// The northeast corner of the rectangle.
-  final LatLng northeast;
-
   /// Creates geographical bounding box with the specified corners.
   ///
   /// The latitude of the southwest corner cannot be larger than the
@@ -74,6 +68,12 @@ class LatLngBounds {
       : assert(southwest != null),
         assert(northeast != null),
         assert(southwest.latitude <= northeast.latitude);
+
+  /// The southwest corner of the rectangle.
+  final LatLng southwest;
+
+  /// The northeast corner of the rectangle.
+  final LatLng northeast;
 
   dynamic _toJson() {
     return <dynamic>[southwest._toJson(), northeast._toJson()];