+++ packages/shared_preferences/example/lib/main.dart
@@ -32,7 +32,7 @@ class SharedPreferencesDemoState extends State<SharedPreferencesDemo> {
   Future<SharedPreferences> _prefs = SharedPreferences.getInstance();
   Future<int> _counter;
 
-  Future<Null> _incrementCounter() async {
+  Future<void> _incrementCounter() async {
     final SharedPreferences prefs = await _prefs;
     final int counter = (prefs.getInt('counter') ?? 0) + 1;