+++ packages/video_player/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.7.2
+
+* Updated to use factories on exoplayer `MediaSource`s for Android instead of the now-deprecated constructors.
+
+## 0.7.1
+
+* Fixed null exception on Android when the video has a width or height of 0.
+
 ## 0.7.0
 
 * Add a unit test for controller and texture changes. This is a breaking change since the interface